@@ -1,6 +1,8 @@
 {
   "name": "metamodels/attribute_translatedcontentarticle",
   "description": "MetaModels attribute for translated content article",
+  "license": "LGPL-3.0-or-later",
+  "type": "contao-bundle",
   "keywords": [
     "contao",
     "metamodels",
@@ -8,9 +10,6 @@
     "article",
     "translatable"
   ],
-  "type": "contao-bundle",
-  "homepage": "https://now.metamodel.me/",
-  "license": "LGPL-3.0-or-later",
   "authors": [
     {
       "name": "Andreas Dziemba",
@@ -19,6 +18,7 @@
       "role": "Developer"
     }
   ],
+  "homepage": "https://now.metamodel.me/",
   "support": {
     "email": "mail@metamodel.me",
     "issues": "https://github.com/MetaModels/attribute_translatedcontentarticle/issues",
@@ -27,35 +27,42 @@
     "source": "https://github.com/MetaModels/attribute_translatedcontentarticle"
   },
   "require": {
-    "php": "^7.4 || ^8.0",
+    "php": "^8.1",
     "ext-pdo": "*",
-    "contao-community-alliance/dc-general": "^2.3@dev",
-    "contao/core-bundle": "^4.13.0, <5.0",
-    "metamodels/core": "^2.3@dev",
+    "contao-community-alliance/dc-general": "^2.3",
+    "contao/core-bundle": "^4.13.0 <5.0",
+    "metamodels/core": "dev-feature/phpcq-2 as 2.3.0",
     "symfony/dependency-injection": "^5.4",
     "symfony/http-kernel": "^5.4"
   },
   "require-dev": {
     "contao/manager-plugin": "^2.1",
     "menatwork/contao-multicolumnwizard-bundle": "^3.4",
-    "phpcq/all-tasks": "^1.2"
+    "phpcq/runner-bootstrap": "^1.0@dev"
   },
+  "repositories": [
+    {
+      "type": "composer",
+      "url": "https://packages.cyberspectrum.de/r"
+    }
+  ],
   "autoload": {
     "psr-4": {
       "MetaModels\\AttributeTranslatedContentArticleBundle\\": "src/"
     }
   },
-  "extra": {
-    "contao-manager-plugin": "MetaModels\\AttributeTranslatedContentArticleBundle\\ContaoManager\\Plugin",
-    "branch-alias": {
-      "dev-feature/2.3.0": "2.3.x-dev"
-    }
-  },
   "config": {
     "allow-plugins": {
       "contao-components/installer": false,
-      "contao/manager-plugin": false
+      "contao/manager-plugin": false,
+      "php-http/discovery": true
     },
     "sort-packages": true
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-feature/2.3.0": "2.3.x-dev"
+    },
+    "contao-manager-plugin": "MetaModels\\AttributeTranslatedContentArticleBundle\\ContaoManager\\Plugin"
   }
 }
