@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_translatedcontentarticle.
  *
- * (c) 2012-2023 The MetaModels team.
+ * (c) 2012-2024 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,7 +15,7 @@
  * @author     Andreas Dziemba <adziemba@web.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
- * @copyright  2012-2023 The MetaModels team.
+ * @copyright  2012-2024 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_translatedcontentarticle/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -78,16 +78,16 @@ class ContentArticleWidget extends AbstractWidget
     /**
      * The database connection.
      *
-     * @var \Doctrine\DBAL\Connection
+     * @var Connection
      */
-    private $connection;
+    private Connection $connection;
 
     /**
      * The contao input.
      *
-     * @var \Contao\CoreBundle\Framework\Adapter|Input
+     * @var Adapter|Input
      */
-    private $input;
+    private Adapter|Input $input;
 
     /**
      * Compat layer.
@@ -116,6 +116,7 @@ class ContentArticleWidget extends AbstractWidget
             );
             // @codingStandardsIgnoreEnd
             $connection = System::getContainer()->get('database_connection');
+            assert($connection instanceof Connection);
         }
         $this->connection = $connection;
 
@@ -126,18 +127,21 @@ class ContentArticleWidget extends AbstractWidget
                 E_USER_DEPRECATED
             );
             // @codingStandardsIgnoreEnd
-            $input = System::getContainer()->get('contao.framework')->getAdapter(Input::class);
+            $input = System::getContainer()->get('contao.framework')?->getAdapter(Input::class);
+            assert($input instanceof Adapter);
         }
         $this->input = $input;
 
-        if (null === $translator) {
+        if (null ===  $translator) {
             // @codingStandardsIgnoreStart
             @trigger_error(
                 'Translator is missing. It has to be passed in the constructor. Fallback will be dropped.',
                 E_USER_DEPRECATED
             );
             // @codingStandardsIgnoreEnd
+
             $translator = System::getContainer()->get('translator');
+            assert($translator instanceof TranslatorInterface);
         }
         $this->translator = $translator;
 
@@ -224,9 +228,10 @@ class ContentArticleWidget extends AbstractWidget
             }
         }
 
-        $rootTable = $this->getRootMetaModelTable($this->strTable);
+        $rootTable    = $this->getRootMetaModelTable($this->strTable);
+        $requestToken = System::getContainer()->get('contao.csrf.token_manager')?->getDefaultTokenValue();
 
-        $strQuery = http_build_query([
+        $strQuery = \http_build_query([
                                          'do'          => 'metamodel_' . ($rootTable ?: 'table_not_found'),
                                          'table'       => 'tl_content',
                                          'ptable'      => $this->strTable,
@@ -237,7 +242,7 @@ class ContentArticleWidget extends AbstractWidget
                                          'popup'       => 1,
                                          'nb'          => 1,
                                          'langSupport' => 1,
-                                         'rt'          => REQUEST_TOKEN,
+                                         'rt'          => $requestToken,
                                      ]);
 
         $contentElements =
@@ -276,7 +281,7 @@ class ContentArticleWidget extends AbstractWidget
             ->select('t.tableName, d.renderType, d.ptable')
             ->from('tl_metamodel', 't')
             ->leftJoin('t', 'tl_metamodel_dca', 'd', '(t.id=d.pid)')
-            ->execute();
+            ->executeQuery();
 
         while ($row = $statement->fetchAssociative()) {
             $tables[$row['tableName']] = [
@@ -343,7 +348,7 @@ class ContentArticleWidget extends AbstractWidget
             ->setParameter('ptable', $ptableName)
             ->setParameter('slot', $slotName)
             ->setParameter('lang', $currentLang)
-            ->execute();
+            ->executeQuery();
 
         while ($row = $statement->fetchAssociative()) {
             $contentElements[] = [
